ZEL_CHATBOT/requirements.txt
langchain==0.0.237
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.22
openai==0.27.6
faiss-cpu==1.7.4
gTTS==2.3.2
simpleplayer==0.0.2


# uncomment to use huggingface llms
 huggingface-hub==0.14.1

# uncomment to use instructor embeddings
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
